+++ com.unity.cinemachine/CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - CinemachineGroupFraming now has a compatibility mode so that it can work with CinemachineConfiner2D out of the box.
 
 ### Added
-- Added Recenter Target setting to CinemachinePanTilt
+- New Rotation Control behaviour SplineDollyLookAtTargets lets you specify LookAt points at desird positions along a Spline Dolly trajectory.
+- Added Recenter Target setting to CinemachinePanTilt.
 
 
 ## [3.1.0] - 2024-04-01